feat: useUpdateFunctionsStore separated into two files
